+++ crates/miden-lib/src/transaction/mod.rs
@@ -142,7 +142,7 @@ impl TransactionKernel {
 
         let stack_inputs = TransactionKernel::build_input_stack(
             account.id(),
-            account.init_commitment(),
+            account.initial_commitment(),
             tx_inputs.input_notes().commitment(),
             tx_inputs.block_header().commitment(),
             tx_inputs.block_header().block_num(),
@@ -204,7 +204,7 @@ impl TransactionKernel {
     /// - INPUT_NOTES_COMMITMENT, see `transaction::api::get_input_notes_commitment`.
     pub fn build_input_stack(
         account_id: AccountId,
-        init_account_commitment: Word,
+        initial_account_commitment: Word,
         input_notes_commitment: Word,
         block_commitment: Word,
         block_num: BlockNumber,
@@ -215,7 +215,7 @@ impl TransactionKernel {
         inputs.push(account_id.suffix());
         inputs.push(account_id.prefix().as_felt());
         inputs.extend(input_notes_commitment);
-        inputs.extend_from_slice(init_account_commitment.as_elements());
+        inputs.extend_from_slice(initial_account_commitment.as_elements());
         inputs.extend_from_slice(block_commitment.as_elements());
         StackInputs::new(inputs)
             .map_err(|e| e.to_string())