+++ crates/miden-objects/src/transaction/tx_witness.rs
@@ -97,7 +97,7 @@ mod tests {
         );
 
         let tx_inputs = TransactionInputs::new(
-            account.clone(),
+            account,
             None,
             block_header.clone(),
             partial_blockchain.clone(),