+++ crates/miden-tx/src/prover/prover_host.rs
@@ -72,7 +72,9 @@ where
     }
 
     /// Consumes `self` and returns the account delta, output notes and transaction progress.
-    pub fn into_parts(self) -> (AccountDelta, Vec<OutputNote>, TransactionProgress) {
+    pub fn into_parts(
+        self,
+    ) -> (AccountDelta, InputNotes<InputNote>, Vec<OutputNote>, TransactionProgress) {
         self.base_host.into_parts()
     }
 }